refactor: merge wshandler constants/helpers into support (#2908)
