+++ src/klangk/klangk/wshandler/__init__.py
@@ -2,8 +2,9 @@
 
 Historically all WebSocket handling lived in a single ~3000-line
 ``wshandler.py``.  That module has been split into per-concern
-submodules (``constants``, ``safe_websocket``, ``session``,
-``controllers``, ``connection``, ``dispatch``, ``helpers``).
+submodules (``support``, ``safe_websocket``, ``session``,
+``window_watcher``, ``controllers``, ``connection``, ``dispatch``,
+``decider``, ``sidecar``).
 
 This package re-exports every public (and the few private) names from
 those submodules so existing call sites keep working unchanged, e.g.::
@@ -22,11 +23,19 @@ from .. import podman as podman  # noqa: F401
 from .. import terminal as terminal  # noqa: F401
 from .. import workspaces as workspaces  # noqa: F401
 
-from .constants import (
+from .support import (
     MAX_INPUT_SIZE as MAX_INPUT_SIZE,
     SEND_QUEUE_SIZE as SEND_QUEUE_SIZE,
     WS_DEBUG as WS_DEBUG,
+    disconnect_all_websockets as disconnect_all_websockets,
+    disconnect_user as disconnect_user,
+    format_container_info as format_container_info,
+    format_idle_timeout as format_idle_timeout,
     log_ws_msg as log_ws_msg,
+    refresh_user_handle as refresh_user_handle,
+    reset_workspace_state as reset_workspace_state,
+    send_error as send_error,
+    send_event as send_event,
 )
 from .safe_websocket import (
     SafeWebSocket as SafeWebSocket,
@@ -37,6 +46,7 @@ from .safe_websocket import (
 from .session import (
     WebSocketState as WebSocketState,
     WorkspaceSession as WorkspaceSession,
+    get_shared_terminals as get_shared_terminals,
 )
 from .controllers import (
     ExecController as ExecController,
@@ -51,15 +61,4 @@ from .dispatch import (
     WS_CONNECTION_COMMANDS as WS_CONNECTION_COMMANDS,
     WS_STATE_COMMANDS as WS_STATE_COMMANDS,
     handle_websocket as handle_websocket,
-)
-from .helpers import (
-    format_container_info as format_container_info,
-    format_idle_timeout as format_idle_timeout,
-    get_shared_terminals as get_shared_terminals,
-    send_event as send_event,
-    disconnect_all_websockets as disconnect_all_websockets,
-    disconnect_user as disconnect_user,
-    refresh_user_handle as refresh_user_handle,
-    reset_workspace_state as reset_workspace_state,
-    send_error as send_error,
 )
